+++ R/f.R
@@ -5,17 +5,17 @@
 #' @return A formula object
 #' @export
 #' @examples
-#' f_new(quote(a))
-#' f_new(quote(a), quote(b))
-f_new <- function(rhs, lhs = NULL, env = parent.frame()) {
+#' new_f(quote(a))
+#' new_f(quote(a), quote(b))
+new_f <- function(rhs, lhs = NULL, env = parent.frame()) {
   if (!is.environment(env) && !is_null(env)) {
     stop("`env` must be an environment", call. = FALSE)
   }
 
   if (is.null(lhs)) {
-    f <- call_new("~", rhs)
+    f <- new_call("~", rhs)
   } else {
-    f <- call_new("~", lhs, rhs)
+    f <- new_call("~", lhs, rhs)
   }
 
   structure(f, class = "formula", .Environment = env)
@@ -52,7 +52,7 @@ f_rhs <- function(f) {
 #' @rdname f_rhs
 `f_rhs<-` <- function(x, value) {
   stopifnot(is_formula(x))
-  f_new(value, f_lhs(x), f_env(x))
+  new_f(value, f_lhs(x), f_env(x))
 }
 
 #' @export
@@ -66,7 +66,7 @@ f_lhs <- function(f) {
 #' @rdname f_rhs
 `f_lhs<-` <- function(x, value) {
   stopifnot(is_formula(x))
-  f_new(f_rhs(x), value, f_env(x))
+  new_f(f_rhs(x), value, f_env(x))
 }
 
 #' @export
@@ -82,7 +82,7 @@ f_env <- function(f) {
 #' @rdname f_rhs
 `f_env<-` <- function(x, value) {
   stopifnot(is_formula(x))
-  f_new(f_rhs(x), f_lhs(x), value)
+  new_f(f_rhs(x), f_lhs(x), value)
 }
 
 #' Turn RHS of formula into a string/label.
@@ -135,7 +135,7 @@ f_unwrap <- function(f) {
   if (identical(e, emptyenv())) {
     f
   } else {
-    f_new(substitute_(f_rhs(f), e), f_lhs(f), parent.env(e))
+    new_f(substitute_(f_rhs(f), e), f_lhs(f), parent.env(e))
   }
 }
 
@@ -208,7 +208,7 @@ as_f_list <- function(x) {
 #' env(f)
 #'
 #' # The user can supply her own environment:
-#' env <- env_new()
+#' env <- new_env()
 #' f <- api_function(quote(foobar), env)
 #' identical(env(f), env)
 #'
@@ -229,6 +229,6 @@ as_fquote <- function(expr, env = NULL) {
     expr
   } else {
     env <- env %||% caller_env(2)
-    f_new(expr, env = env)
+    new_f(expr, env = env)
   }
 }