+++ R/vector.R
@@ -224,55 +224,55 @@ rep_along <- function(.x, .y, ..., .attrs = list()) {
 #' Create new vectors.
 #'
 #' These functions construct vectors of given length, with attributes
-#' specified via dots. Except for \code{lst_new()} and
-#' \code{raw_new()}, the empty vectors are filled with typed
+#' specified via dots. Except for \code{new_lst()} and
+#' \code{new_raw()}, the empty vectors are filled with typed
 #' \link{missing} values. This is in contrast to the base function
 #' \code{\link[base]{vector}()} which creates zero-filled vectors.
 #'
 #' @inheritParams with_attributes
 #' @param .n The vector length.
 #' @examples
-#' lst_new(10)
+#' new_lst(10)
 #'
 #' # Add attributes, including the S3 class:
-#' int_new(0, index = 1)
-#' dbl_new(10, class = "my_class")
+#' new_int(0, index = 1)
+#' new_dbl(10, class = "my_class")
 #' @name new-vectors
 #' @seealso along
 NULL
 
 #' @export
 #' @rdname new-vectors
-lgl_new <- function(.n = 0, ..., .attrs = list()) {
+new_lgl <- function(.n = 0, ..., .attrs = list()) {
   with_attributes(rep_len(na_lgl, .n), ..., .attrs = .attrs)
 }
 #' @export
 #' @rdname new-vectors
-int_new <- function(.n = 0, ..., .attrs = list()) {
+new_int <- function(.n = 0, ..., .attrs = list()) {
   with_attributes(rep_len(na_int, .n), ..., .attrs = .attrs)
 }
 #' @export
 #' @rdname new-vectors
-dbl_new <- function(.n = 0, ..., .attrs = list()) {
+new_dbl <- function(.n = 0, ..., .attrs = list()) {
   with_attributes(rep_len(na_dbl, .n), ..., .attrs = .attrs)
 }
 #' @export
 #' @rdname new-vectors
-chr_new <- function(.n = 0, ..., .attrs = list()) {
+new_chr <- function(.n = 0, ..., .attrs = list()) {
   with_attributes(rep_len(na_chr, .n), ..., .attrs = .attrs)
 }
 #' @export
 #' @rdname new-vectors
-lst_new <- function(.n = 0, ..., .attrs = list()) {
+new_lst <- function(.n = 0, ..., .attrs = list()) {
   with_attributes(vector("list", .n), ..., .attrs = .attrs)
 }
 #' @export
 #' @rdname new-vectors
-cpl_new <- function(.n = 0, ..., .attrs = list()) {
+new_cpl <- function(.n = 0, ..., .attrs = list()) {
   with_attributes(rep_len(na_cpl, .n), ..., .attrs = .attrs)
 }
 #' @export
 #' @rdname new-vectors
-raw_new <- function(.n = 0, ..., .attrs = list()) {
+new_raw <- function(.n = 0, ..., .attrs = list()) {
   with_attributes(vector("raw", .n), ..., .attrs = .attrs)
 }